requirements.txt
beautifulsoup4==4.6.0
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.6
coreschema==0.0.4
dj-database-url==0.4.1
Django==1.11.4
django-filter==1.0.4
djangorestframework==3.7.0
gunicorn==19.6.0
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
pytz==2017.2
requests==2.14.2
simplejson==3.11.1
uritemplate==3.0.0
virtualenv==15.1.0
virtualenvwrapper-win==1.2.1
whitenoise==3.2
